chore: mirror test layout to src structure for outlook, relocate misfiled tests
Reorganize outlook tests to match src folder structure with mirrored hierarchy:
- adapters/ for graph-* and adapter tests (graph-mailbox-connector-* → mailbox-connector-*; graph-attachment-fetch → mailbox-connector-attachment-fetch)
- services/backup/ for mailbox sync and attachment tests (mailbox-sync-* → *; drop prefixes where folder provides context)
- services/restore/ for restore-related tests (drop .helper suffix to match renamed src)
- services/save/ for eml/zip/integrity tests
- services/shared/ for mime/folder-selector tests
- services/status/ for status tests
- services/operation-pre-abort.test.ts spans multiple layers

Unify fixture naming to .fixtures.ts convention:
- mailbox-sync-test-fixtures → mailbox-sync.fixtures
- graph-mailbox-connector.harness → mailbox-connector.fixtures
- restore-service-test-fixtures → restore.service.fixtures (fix cross-package relative import)

Relocate misfiled tests:
- restore-merge.helper.test.ts (core module) → core/tests/unit/services/snapshot-merge.test.ts
- subsite-enumerator.test.ts (adapter) → sharepoint/tests/unit/adapters/
- drive-metadata-capture.test.ts (adapter) → onedrive/tests/unit/adapters/

Drive package tests (onedrive, sharepoint) deferred per issue requirement (src reshuffle in #151/#153).

Fixes #158
