+++ packages/core/tests/unit/services/snapshot-merge.test.ts
