+++ packages/outlook/tests/unit/adapters/attachment-type-mapping.test.ts
