+++ packages/outlook/tests/unit/adapters/mailbox-connector-immutable-id.test.ts
@@ -1,8 +1,8 @@
 import 'reflect-metadata';
 import { describe, it, expect, beforeEach } from 'vitest';
 import type { GraphMailboxConnector } from '@/adapters/graph-mailbox-connector.adapter';
-import type { MockClient } from './graph-mailbox-connector.harness';
-import { create_mock_client, create_connector } from './graph-mailbox-connector.harness';
+import type { MockClient } from './mailbox-connector.fixtures';
+import { create_mock_client, create_connector } from './mailbox-connector.fixtures';
 
 // Issue #48: every Graph request that produces or consumes a message/folder ID
 // must carry Prefer: IdType="ImmutableId". Mixing ID formats corrupts
